wip: split up ui settings
