+++ web/pages/Settings/ui/common.ts
@@ -1,0 +1,24 @@
+import { UI } from '/common/types'
+
+const msgInlineLabels: Record<UI.MessageOption, string> = {
+  edit: 'Edit',
+  regen: 'Regenerate',
+  prompt: 'Prompt View',
+  fork: 'Fork',
+  trash: 'Delete',
+  attach: 'Attach',
+  'schema-regen': 'Retry Schema',
+  visible: 'Visibility',
+  'gen-image': 'Gen Image',
+}
+
+export function toInlineList(opts: UI.UISettings['msgOptsInline']) {
+  const next = Object.entries(opts).map(([key, item], i) => ({
+    id: i,
+    value: key,
+    label: msgInlineLabels[key as UI.MessageOption],
+    enabled: item.outer,
+  }))
+
+  return next
+}
