+++ web/variables.css
@@ -1,5 +1,6 @@
 :root {
   --sitewide-font: unset;
+  --sitewide-font-size: 16px;
 
   --text-bot-background: unset;
   --text-chatcolor: unset;
